+++ packages/spec/examples/common/index.js
@@ -1,2 +1,5 @@
 const foo = [1, 2];
 console.log(foo);
+
+const t = { s: 1 };
+console.log(t['s']);