+++ packages/spec/examples/common-ts/index.ts
@@ -5,3 +5,6 @@ interface AppDeveloper {
 
 const bar = [1, 2];
 const one: AppDeveloper = { name: 'hello', id: 1 };
+
+const t = { s: 1 };
+console.log(t['s']);