+++ packages/spec/src/eslint/common.js
@@ -6,9 +6,10 @@ module.exports = {
   ],
   rules: {
     // Change error to warn
-    semi: 'warn',
+    'semi': 'warn',
     'eol-last': 'warn',
     'quote-props': 'warn',
     'no-unused-vars': 'warn',
+    'dot-notation': 'off',
   },
 };